Update to latest jeweler, split associations into singletons/has many
Jeweler now supports singleton/has many state based on the association, instead of class level. This solves a lot of problems where the same object would pose as singleton in one context and as has many association in the next. In order to keep the current Keyline gem’s interface and not break anything, I did not cleanup classes that are now obsolete because of this change. This will follow in a later commit.