Add support for retrieving some invoice data and documents
I added accounting API support at this point to be able to retrieve the PDFs, thus the accounting objects are not writable yet and no associations are configured (line items etc.). The Keyline API is already capable of writing and will also provide all child objects, so this can be added to the gem at a later time when needed.